cloud/gateway_api/requirements.txt
Flask==2.1.2
gunicorn==20.1.0

google-cloud-pubsub==2.12.0
google-cloud-datastore==2.5.1